// finance_client/src/app/user/data/types/sub-category.ts
export interface SubCategory {
  category_id?: number;
  deletedBy?: string;
  deletedFlag?: string;
  deletedTime?: Date;
  glSubHead?: string
  id?: number;
  modifiedBy?: string;
  modifiedFlag?: string;
  modifiedTime?: Date
  postedFlag?: string;
  postedTime?: Date;
  postedBy?: DOMException,
  reason?: string;
  status?: string;
  subcategoryDescription?: string;
  subcategoryName?: string;
  verifiedBy?: string;
  verifiedFlag?: string;
  verifiedTime?: Date
}